+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog  
 
+## 10.1.0
+### New Feature
+* The `source` attribute within the installedApks object is now returned in the API(TerminalApi-`getTerminal`, `getTerminalBySn`) response.
+
+
 ## 10.0.0
 ### New Feature
 * Add parameter query `includeMasterTerminal` and return in TerminalApi-`getTerminal`, `getTerminalBySn` API